+++ src/Products/PloneMeeting/vocabularies.py
@@ -940,6 +940,30 @@ class HasAnnexesToSignVocabulary(object):
 HasAnnexesToSignVocabularyFactory = HasAnnexesToSignVocabulary()
 
 
+class BooleanForFacetedVocabulary(object):
+    implements(IVocabularyFactory)
+
+    def __call__(self, context):
+        """ """
+        res = []
+        res.append(SimpleTerm('0',
+                              '0',
+                              safe_unicode(translate('boolean_value_false',
+                                                     domain='PloneMeeting',
+                                                     context=context.REQUEST)))
+                   )
+        res.append(SimpleTerm('1',
+                              '1',
+                              safe_unicode(translate('boolean_value_true',
+                                                     domain='PloneMeeting',
+                                                     context=context.REQUEST)))
+                   )
+        return SimpleVocabulary(res)
+
+
+BooleanForFacetedVocabularyFactory = BooleanForFacetedVocabulary()
+
+
 class DownOrUpWorkflowAgainVocabulary(object):
     implements(IVocabularyFactory)